\name{ENVISIONQuery-package}
\title{ENVISIONQuery.}
\description{ENVISIONQuery.}
\details{\tabular{ll}{
Package: \tab ENVISIONQuery\cr
Type: \tab Package\cr
Version: \tab 1.1.0 \cr
Date: \tab 2011-04-19 \cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}}
\alias{ENVISIONQuery-package}
\docType{package}
\author{Alex Lisovich, Roger Day}
\keyword{package}
