\name{getFormatter}
\alias{getFormatter}
\title{Get the formatting function for a given tool...}
\usage{getFormatter(tool)}
\description{Get the formatting function for a given tool}
\value{The formatting function for a given service.}
\keyword{internal}
\author{Alex Lisovich, Roger Day}
\arguments{\item{tool}{Tool handler}}
