\name{getMerger}
\alias{getMerger}
\title{Get the merging function for a given tool.}
\usage{getMerger(tool)}
\description{Get the merging function for a given tool.}
\details{If possible, meerging function merges the list of formatted query result chunks into a single formatted object (typically a data frame).}
\value{The merging function for a given service.}
\keyword{internal}
\author{Alex Lisovich, Roger Day}
\arguments{\item{tool}{Tool handler}}
