\name{getServiceNames}
\alias{getServiceNames}
\title{Get the names of available Envision services...}
\description{Get the names of available Envision services}
\value{Character vector representing the list of of available Envision service names.}
\author{Alex Lisovich, Roger Day}
\examples{#get the names of all available services
serviceNames<-getServiceNames();
print(serviceNames);}
