\name{getServices}
\alias{getServices}
\title{Get available services...}
\description{Get available services}
\value{The multi-level list of services and service attributes}
\author{Alex Lisovich, Roger Day}
\examples{#get the handlers for all available services including utilities
services<-getServices();
names(services);}
