\name{jgetOptions}
\alias{jgetOptions}
\title{Convert options list into the character vector.}
\usage{jgetOptions(service, options)}
\description{Convert options list into the character vector.}
\details{Convert options list into the character vector suitable for submission
to Java Web Services subsystem.}
\value{Character vector containing set of <option name,value> pairs}
\keyword{internal}
\author{Alex Lisovich, Roger Day}
\arguments{\item{service}{Service handler}
\item{options}{Options list}}
