\name{mergeFrameList}
\alias{mergeFrameList}
\title{Merges the data frames representing the partial query results into a single data frame.}
\usage{mergeFrameList(resList, verbose=FALSE)}
\description{Merges the data frames representing the partial query results into a single data frame.}
\value{A single data frame representing the merged query results.}
\keyword{internal}
\author{Alex Lisovich, Roger Day}
\arguments{\item{resList}{List of data frames.}
\item{verbose}{if TRUE enables diagnostic messages. Default is FALSE.}}
