\name{progressMsg}
\alias{progressMsg}
\title{Display the task progress message.}
\usage{progressMsg(msg, i, total)}
\description{Display the task progress message.}
\details{Displays the percentage of task completeness accompanied by the custom message by repatedly updating the same output line.}
\keyword{internal}
\author{Alex Lisovich, Roger Day}
\arguments{\item{msg}{Custom message to be displayed.}
\item{i}{Current task step.}
\item{total}{Total number of task steps.}}
